# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bauk Cheescake",
        "version": "0.0.11",
        "description": "appiid: 3232f21e-653e-4e98-b46d-c34cbc751182 - TODO: All users can post new addresses"
    },
    "servers": [
        {
            "url": "https://localhost:443"
        }
    ],
    "security": [
        {
            "BearerAuth": [],
            "Refresh": [],
            "MailAuth": [],
            "AppiId": []
        }
    ],
    "paths": {
...
        "/additions/": {
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "description": "Get additions",
                "tags": [
                    "Additions"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "type": "array",
                                    "items":{
                                        "$ref": "#/components/schemas/Addition"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Addition"
                            }
                        }
                    }
                },
                "description": "Post additions",
                "tags": [
                    "Additions"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/additions/{id}": {
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get addition by id",
                "tags": [
                    "Additions"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Addition"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Addition"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Edit addition",
                "tags": [
                    "Additions"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/additions/multi": {
            "post": {
                "security": [
                    {
                        "AppiId": [],
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Addition"
                                }
                            }
                        }
                    }
                },
                "description": "Post additions",
                "tags": [
                    "Additions"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "Refresh": {
                "type": "apiKey",
                "in": "header",
                "name": "Refresh"
            },
            "MailAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "MailToken"
            },
            "AppiId": {
                "type": "apiKey",
                "in": "header",
                "name": "appiId"
            }
        },
        "schemas": {
...
            "Addition": {
                "type": "object",
                "properties": {
                    "Name": {
                        "type": "string"
                    },
                    "Price": {
                        "type": "integer"
                    }
                }
            },
...
        }
    }
}